rppg-mmtcan/pyproject.toml
[project]
name = "rppg-mmtcan"
version = "0.1.0"
description = ""
authors = [
    {name = "djs manoj",email = "djsmanoj0000@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0,<3.0.0",
    "torchvision>=0.17.0,<0.18.0",
    "numpy (>=1.24.0,<2.0.0)",
    "facenet-pytorch (==2.6.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.1,<0.35.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "opencv-python-headless (>=4.11.0.86,<5.0.0.0)",
    "scipy (>=1.15.2,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "rppg-mmtcan"
version = "0.1.0"
description = "A project for rPPG signal extraction using MMT-CAN."
authors = ["Your Name <your.email@example.com>"]
packages = [{ include = "src" }]
